requirements.txt
flask
python-dotenv
openai
requests
Flask-SQLAlchemy
flask-login
Flask-Migrate